+++ R/a18_c_H2O.R
@@ -131,11 +131,11 @@
 #'
 #' Kim, S.-T., Mucci, A., & Taylor, B. E. (2007).
 #' Phosphoric acid fractionation factors for calcite and aragonite between
-#' 25 and 75 °C: Revisited. Chemical Geology, 246(3-4), 135-146.
+#' 25 and 75 °C: Revisited. Chemical Geology, 246(3-4), 135-146.
 #' \doi{10.1016/j.chemgeo.2007.08.005}
 #'
 #' Coplen, T. B. (2007).
-#' Calibration of the calcite–water oxygen-isotope geothermometer
+#' Calibration of the calcite-water oxygen-isotope geothermometer
 #' at Devils Hole, Nevada, a natural laboratory.
 #' Geochimica et Cosmochimica Acta, 71(16), 3948-3957.
 #' \doi{10.1016/j.gca.2007.05.028}
@@ -144,7 +144,7 @@
 #' Amiot, R., et al. (2010).
 #' Oxygen isotope fractionation between apatite-bound carbonate
 #' and water determined from controlled experiments with synthetic
-#' apatites precipitated at 10–37°C.
+#' apatites precipitated at 10-37°C.
 #' Geochimica et Cosmochimica Acta, 74(7), 2072-2081.
 #' \doi{10.1016/j.gca.2009.12.024}
 #'
@@ -178,7 +178,7 @@
 #' Benning, L.G. & Bernasconi, S.M. (2019).
 #' Calibration of the oxygen and clumped isotope thermometers for
 #' (proto-)dolomite based on synthetic and natural carbonates.
-#' Chemical Geology, 525, 1–17.
+#' Chemical Geology, 525, 1-17.
 #' \doi{10.1016/j.chemgeo.2019.07.014}
 #'
 #' @family fractionation_factors
@@ -199,10 +199,10 @@ a18_c_H2O = function(temp, min, eq) {
       # Coplen (2007)
       exp((17.4 * 1000 / TinK - 28.6) / 1000)
     } else if (eq == "KO97-orig") {
-      # Kim and O'Neil (1997) –– original
+      # Kim and O'Neil (1997) -- original
       exp((18.03 * 1000 / TinK - 32.42) / 1000)
     } else if (eq == "KO97") {
-      # Kim and O'Neil (1997) –– reprocessed
+      # Kim and O'Neil (1997) -- reprocessed
       exp((18.04 * 1000 / TinK - 32.18) / 1000)
     } else if (eq == "Watkins13") {
       # Watkins et al. (2013)